fix(Makefile): Build fpga software before bitstreaem.
Since the current fpga makefile build process depends on the software,
then it must be built first.
And since the software build process requires the riscv-gnu-toolchain,
then it must run in the nix-shell environment.
However, the FPGA build process with vivado does not behave well inside
a nix-shell environment.
So these two processes must be run separately.

The `fpga-run` makefile target will also re-build the FPGA software again, so this is not the most efficient solution.